refactor: improve import statements and logging in various modules; update TypeScript configuration
