+++ src/analyzer/callgraph/GraphExtractor.ts
@@ -24,6 +24,7 @@
 
 import { WasmParserFactory } from "@/analyzer/languages/WasmParserFactory";
 import type { RelationType, SupportedLang, SymbolType } from "@/shared/callgraph-types";
+import { getLogger } from "@/shared/logger";
 import { normalizePath } from "@/shared/path";
 import fs from "node:fs/promises";
 import path from "node:path";
@@ -438,7 +439,7 @@ export class GraphExtractor {
       this.querySourceCache.set(normalizedLang, src);
       return src;
     } catch (err) {
-      console.error(`[CallGraph] Query file not found: ${queryPath}`, err);
+      getLogger('GraphExtractor').error(`Query file not found: ${queryPath}`, err);
       this.querySourceCache.set(normalizedLang, "");
       return "";
     }
