--- src/webview/components/reactflow/cycles.ts
@@ -1,8 +1,0 @@
-/**
- * Cycle detection (DFS) for directed graphs.
- * Returns a set of node ids involved in a cycle.
- *
- * @deprecated Import directly from '@/analyzer/callgraph/cycleUtils' instead.
- * This file re-exports for backward compatibility.
- */
-export { detectCycles } from '../../../analyzer/callgraph/cycleUtils';
